+++ R/mod_qtl_view.R
@@ -343,7 +343,8 @@ mod_qtl_view_server <- function(input, output, session,
     )
     
     parents <- unlist(strsplit(input$parents_name, ","))
-    
+    parents <- gsub(" ", "", parents)
+
     withProgress(message = 'Working:', value = 0, {
       incProgress(0.5, detail = paste("Getting data..."))
       data <- data_effects(qtl_info = loadQTL()$qtl_info, 
@@ -477,7 +478,11 @@ mod_qtl_view_server <- function(input, output, session,
       need(all(input$haplo != "Select `bar` design to access this feature."), "Select `bar` design to access this feature.")
     )
     
-    list.p <- select_haplo(input$haplo, loadQTL()$probs, loadQTL()$selected_mks, effects.data(), exclude.haplo = input$haplo_exclude)
+    list.p <- select_haplo(input.haplo = as.list(input$haplo), 
+                           exclude.haplo = as.list(input$haplo_exclude),
+                           probs = loadQTL()$probs, 
+                           selected_mks = loadQTL()$selected_mks, 
+                           effects.data = effects.data())
     p <- list.p[[1]]
     inds <- list.p[[2]]
     counts <- ceiling(length(p)/3)
@@ -620,11 +625,9 @@ mod_qtl_view_server <- function(input, output, session,
     
     df <- brushedPoints(qtl.data()[[2]], input$plot_brush, xvar = "x", yvar = "y.dat")
     
-    if(!grepl("Example" ,input$parents_name)) {
-      cat("here2")
-      parents <- unlist(strsplit(input$parents_name, ","))
-    } else parents <- NULL
-    
+    parents <- unlist(strsplit(input$parents_name, ","))
+    parents <- gsub(" ", "", parents)
+
     data <- data_effects(qtl_info = loadQTL()$qtl_info, 
                          effects = loadQTL()$effects,
                          pheno.col = as.character(df$Trait),