+++ R/mod_upload.R
@@ -7,6 +7,7 @@
 #' @noRd 
 #' 
 #' @importFrom shinyjs inlineCSS useShinyjs
+#' @importFrom hidecan get_example_data GWAS_data_from_gwaspoly GWAS_data DE_data CAN_data
 #' @importFrom shiny NS tagList 
 mod_upload_ui <- function(id){
   ns <- NS(id)
@@ -26,7 +27,7 @@ mod_upload_ui <- function(id){
       column(width = 12,
              fluidPage(
                box(id= ns("box_example"), width = 12, solidHeader = TRUE, collapsible = TRUE, collapsed = FALSE, status="primary", title = actionLink(inputId = ns("exampleID"), label = tags$b("Available example datasets")),
-                   radioButtons(ns("example_map"), label = p("They contain the entire linkage map and QTL analysis but just a subset of individuals."), 
+                   radioButtons(ns("example_map"), label = p("They contain a results for a subset of tetraploid potato linkage map and QTL analysis."), 
                                 choices = c("Potato - Atlantic x B1829-5" = "tetra_map"),
                                 selected = "tetra_map"), br(), br(), hr(),
                    tags$p("Access complete example datasets ", 
@@ -83,6 +84,20 @@ mod_upload_ui <- function(id){
                        ), br(),
                        fileInput(ns("polymapR.dataset"), label = h6("File: polymapR.dataset.RData"), multiple = F),
                        fileInput(ns("polymapR.map"), label = h6("File: polymapR.map.RData"), multiple = F),
+                   ),
+                   box(id = ns("box_onemap"),width = 12, solidHeader = FALSE, collapsible = TRUE, collapsed = TRUE, status="primary",  title = actionLink(inputId = ns("onemapID"), label = tags$b("Upload OneMap output")),
+                       tags$p("Access further information about how to build a linkage maps for diploid outcrossing populations with OneMap ", 
+                              tags$a(href= "https://cristianetaniguti.github.io/Tutorials/onemap/vignettes_highres/Outcrossing_Populations.html","here")), br(),
+                       tags$p("Access further information about how to build a linkage maps for diploid inbred based populations with OneMap ", 
+                              tags$a(href= "https://cristianetaniguti.github.io/Tutorials/onemap/vignettes_highres/Inbred_Based_Populations.html","here")), br(),
+                       tags$p("Access a example code of how to obtain these inputs using OneMap functions ", 
+                              tags$a(href= "https://cristianetaniguti.github.io/viewpoly_vignettes/VIEWpoly_tutorial.html#Upload_linkage_map_files","here")),
+                       hr(),
+                       div(style = "position:absolute;right:1em;",
+                           actionBttn(ns("submit_onemap"), style = "jelly", color = "royal",  size = "sm", label = "submit OneMap", icon = icon("share-square", verify_fa = FALSE)), 
+                       ), br(), br(),
+                       tags$p("Object of class `viewmap`."), 
+                       fileInput(ns("onemap_in"), label = h6("File: my_onemap_map.RData"), multiple = F),
                    ),
                    box(id = ns("box_mapst"), width = 12, solidHeader = FALSE, collapsible = TRUE, collapsed = TRUE, status="primary",title = actionLink(inputId = ns("mapstID"), label = tags$b("Upload linkage map files with standard format (.csv, .tsv or .tsv.gz)")),
                        div(style = "position:absolute;right:1em;",
@@ -234,6 +249,33 @@ mod_upload_ui <- function(id){
       ),
       column(width = 12,
              fluidPage(
+               box(id = ns("box_hidecan"),width = 12, solidHeader = TRUE, collapsible = TRUE, collapsed = TRUE, status="primary", title = actionLink(inputId = ns("hidecanID"), label = tags$b("Upload Hidecan files")),
+                   tags$p("Access further information about the files expected in this section ", 
+                          tags$a(href= "https://plantandfoodresearch.github.io/hidecan/","here")), br(),
+                   
+                   div(style = "position:absolute;right:1em;",
+                       actionBttn(ns("reset_hidecan"), style = "jelly", color = "royal",  size = "sm", label = "reset", icon = icon("undo-alt", verify_fa = FALSE)),
+                       actionBttn(ns("submit_hidecan"), style = "jelly", color = "royal",  size = "sm", label = "submit HIDECAN", icon = icon("share-square", verify_fa = FALSE)), 
+                   ), br(), br(), 
+                   box(id= ns("box_gwaspoly"), width = 12, solidHeader = FALSE, collapsible = TRUE, collapsed = TRUE,  status="primary", title = actionLink(inputId = ns("gwaspolyID"), label = tags$b("Upload GWAS output")),
+                       div(style = "position:absolute;right:1em;",
+                       ), br(), br(),
+                       p("Object of class GWASpoly.thresh obtained with the GWASpoly::set.threshold():"), br(),
+                       fileInput(ns("gwaspoly"), label = h6("File: gwaspoly_res_thr.rda"), multiple = F),
+                       p("or"),
+                       fileInput(ns("gwas"), label = h6("File: gwas.csv"), multiple = T)
+                   ),
+                   box(id= ns("box_gwas_de"), width = 12, solidHeader = FALSE, collapsible = TRUE, collapsed = TRUE,  status="primary", title = actionLink(inputId = ns("gwasID"), label = tags$b("Upload differential expression (DE) and candidate genes (CAN) files")),
+                       div(style = "position:absolute;right:1em;",
+                       ), br(), br(),
+                       fileInput(ns("de"), label = h6("File: DE.csv"), multiple = T),
+                       fileInput(ns("can"), label = h6("File: CAN.csv"), multiple = T)
+                   )
+               )
+             )
+      ),
+      column(width = 12,
+             fluidPage(
                box(id = ns("box_viewpoly"),width = 12, solidHeader = TRUE, collapsible = TRUE, collapsed = TRUE, status="info", title = actionLink(inputId = ns("viewpolyID"), label = tags$b("Download VIEWpoly dataset")),
                    p("The uploaded data are converted to the viewpoly format. It keeps the map and the QTL information. Genome information is not stored."), br(),
                    textInput(ns("data.name"), label = p("Define the dataset name. Do not use spaces between words."), value = "dataset_name"), br(),
@@ -293,6 +335,10 @@ mod_upload_server <- function(input, output, session, parent_session){
     js$collapse(ns("box_mappoly"))
   })
   
+  observeEvent(input$onemapID, {
+    js$collapse(ns("box_onemap"))
+  })
+  
   observeEvent(input$polymapID, {
     js$collapse(ns("box_polymap"))
   })
@@ -323,6 +369,10 @@ mod_upload_server <- function(input, output, session, parent_session){
   
   observeEvent(input$genomeID, {
     js$collapse(ns("box_genome"))
+  })
+  
+  observeEvent(input$hidecanID, {
+    js$collapse(ns("box_hidecan"))
   })
   
   observeEvent(input$viewpolyID, {
@@ -404,6 +454,7 @@ mod_upload_server <- function(input, output, session, parent_session){
   values <- reactiveValues(
     upload_state_map = 0,
     upload_state_mappoly = 0,
+    upload_state_onemap = 0,
     upload_state_polymapR = 0,
     upload_state_map_custom = 0,
     upload_state_qtl = 0,
@@ -411,13 +462,15 @@ mod_upload_server <- function(input, output, session, parent_session){
     upload_state_diaQTL = 0,
     upload_state_polyqtlR = 0,
     upload_state_qtl_custom = 0,
-    upload_state_genome = 0
+    upload_state_genome = 0,
+    upload_state_hidecan = 0
   )
   
   observeEvent(input$reset_all, {
     values$upload_state_viewpoly <- 'reset'
     values$upload_state_map <- 'reset'
     values$upload_state_mappoly = 0
+    values$upload_state_onemap = 0
     values$upload_state_polymapR = 0
     values$upload_state_map_custom = 0
     values$upload_state_qtl <- 'reset'
@@ -426,6 +479,7 @@ mod_upload_server <- function(input, output, session, parent_session){
     values$upload_state_polyqtlR = 0
     values$upload_state_qtl_custom = 0
     values$upload_state_genome <- 'reset'
+    values$upload_state_hidecan <- 'reset'
   })
   
   observeEvent(input$reset_viewpoly, {
@@ -435,6 +489,7 @@ mod_upload_server <- function(input, output, session, parent_session){
   observeEvent(input$reset_map, {
     values$upload_state_map <- 'reset'
     values$upload_state_mappoly = 0
+    values$upload_state_onemap = 0
     values$upload_state_polymapR = 0
     values$upload_state_map_custom = 0
   })
@@ -451,12 +506,21 @@ mod_upload_server <- function(input, output, session, parent_session){
     values$upload_state_genome <- 'reset'
   })
   
+  observeEvent(input$reset_hidecan, {
+    values$upload_state_hidecan <- 'reset'
+  })
+  
   observeEvent(input$submit_viewpoly, {
     values$upload_state_viewpoly <- 'uploaded'
   })
   
   observeEvent(input$submit_mappoly, {
     values$upload_state_mappoly <- 'uploaded'
+    values$upload_state_map <- 0
+  })
+  
+  observeEvent(input$submit_onemap, {
+    values$upload_state_onemap <- 'uploaded'
     values$upload_state_map <- 0
   })
   
@@ -494,9 +558,14 @@ mod_upload_server <- function(input, output, session, parent_session){
     values$upload_state_genome <- 'uploaded'
   })
   
+  observeEvent(input$submit_hidecan, {
+    values$upload_state_hidecan <- 'uploaded'
+  })
+  
   input_map <- reactive({
     if (values$upload_state_map == 0 & 
         values$upload_state_mappoly == 0 & 
+        values$upload_state_onemap == 0 & 
         values$upload_state_polymapR == 0 &
         values$upload_state_map_custom == 0) {
       return(NULL)
@@ -507,6 +576,11 @@ mod_upload_server <- function(input, output, session, parent_session){
         need(!is.null(input$mappoly_in), "Upload mappoly file before submit")
       )
       return(list(mappoly_in = input$mappoly_in))
+    } else if(values$upload_state_onemap == "uploaded"){
+      validate(
+        need(!is.null(input$onemap_in), "Upload onemap file before submit")
+      )
+      return(list(onemap_in = input$onemap_in))
     } else if(values$upload_state_polymapR == "uploaded"){
       validate(
         need(!is.null(input$polymapR.dataset), "Upload polymapR dataset file before submit"),
@@ -618,10 +692,34 @@ mod_upload_server <- function(input, output, session, parent_session){
     })
   })
   
+  input_hidecan <- reactive({
+    if (values$upload_state_hidecan == 0) {
+      return(NULL)
+    } else if (values$upload_state_hidecan == 'reset') {
+      return(NULL)
+    } else if(values$upload_state_hidecan == "uploaded"){
+      validate(
+        need(!all(c(is.null(input$gwas),is.null(input$gwaspoly))), "Upload GWAS results file before submit")
+      )
+      if(!is.null(input$gwaspoly)) {
+        temp <- load(input$gwaspoly$datapath)
+        gwaspoly <- get(temp)
+        gwaspoly <- GWAS_data_from_gwaspoly(gwaspoly)
+        
+      } else gwaspoly <- NULL
+      
+      return(list(GWASpoly = gwaspoly,
+                  GWAS = {if(!is.null(input$gwas)) read_input_hidecan(input$gwas, GWAS_data) else list()},
+                  DE = {if(!is.null(input$de)) read_input_hidecan(input$de, DE_data) else list()},
+                  CAN = {if(!is.null(input$can)) read_input_hidecan(input$can, CAN_data) else list()}))
+    } 
+  })
+  
   # Wait system for the uploads
   loadExample = reactive({
     if(is.null(input_map()$dosages) & is.null(input_map()$phases) & is.null(input_map()$genetic_map) &
        is.null(input_map()$mappoly_in) &
+       is.null(input_map()$onemap_in) &
        is.null(input_map()$polymapR.dataset) &
        is.null(input_map()$polymapR.map) &
        is.null(input_qtl()$selected_mks) & 
@@ -659,6 +757,20 @@ mod_upload_server <- function(input, output, session, parent_session){
       incProgress(0.5, detail = paste("Uploading example map data..."))
       prepare_examples(input$example_map)
     })
+    else NULL
+  })
+  
+  # Load hidecan example
+  loadHidecanExample = reactive({
+    if(is.null(input_hidecan()$gwas) & is.null(input_hidecan()$de) & is.null(input_hidecan()$can))
+      withProgress(message = 'Working:', value = 0, {
+        incProgress(0.5, detail = paste("Uploading example map data..."))        
+        x <- get_example_data()
+        
+        list("GWAS" = list(GWAS_data(x[["GWAS"]])),
+             "DE" = list(DE_data(x[["DE"]])),
+             "CAN" = list(CAN_data(x[["CAN"]])))
+      })
     else NULL
   })
   
@@ -702,6 +814,18 @@ mod_upload_server <- function(input, output, session, parent_session){
       withProgress(message = 'Working:', value = 0, {
         incProgress(0.3, detail = paste("Uploading MAPpoly data..."))
         prepare_MAPpoly(input_map()$mappoly_in)
+      })
+    } else NULL
+  })
+  
+  loadMap_onemap =  reactive({
+    
+    if(!is.null(input_map()$onemap_in)){
+      withProgress(message = 'Working:', value = 0, {
+        incProgress(0.3, detail = paste("Uploading OneMap data..."))
+        temp <- load(input_map()$onemap_in$datapath)
+        viewmap <- get(temp)
+        viewmap
       })
     } else NULL
   })
@@ -810,10 +934,12 @@ mod_upload_server <- function(input, output, session, parent_session){
       if(!is.null(input_genome()$fasta) & !is.null(loadMap())){
         # keep fasta name
         for(i in 1:length(input_genome()$fasta$datapath)){
+          print(file.path(temp_dir(), input_genome()$fasta$name))
+          
           file.rename(input_genome()$fasta$datapath[i], 
                       file.path(temp_dir(), input_genome()$fasta$name[i]))
         }
-        file.path(temp_dir(), input_genome()$fasta$name[1]) 
+        file.path(temp_dir(), sort(input_genome()$fasta$name)[1]) 
       } else if(!is.null(input_genome()$fasta_server) & !is.null(loadMap())) {
         input_genome()$fasta_server
       } else if(!is.null(input_genome()$fasta) | !is.null(input_genome()$fasta_server)) {
@@ -890,6 +1016,7 @@ mod_upload_server <- function(input, output, session, parent_session){
     if(is.null(loadExample()) & 
        is.null(loadMap_custom()) & 
        is.null(loadMap_mappoly()) &
+       is.null(loadMap_onemap()) &
        is.null(loadMap_polymapR()) &
        is.null(loadViewpoly())){
       warning("Select one of the options in `upload` session")
@@ -900,6 +1027,8 @@ mod_upload_server <- function(input, output, session, parent_session){
       return(loadMap_custom())
     } else if(!is.null(loadMap_mappoly())){
       return(loadMap_mappoly())
+    } else if(!is.null(loadMap_onemap())){
+      return(loadMap_onemap())
     } else if(!is.null(loadMap_polymapR())){
       return(loadMap_polymapR())
     } else if(!is.null(loadExample())){
@@ -931,6 +1060,21 @@ mod_upload_server <- function(input, output, session, parent_session){
     }
   })
   
+  loadHidecan = reactive({
+    if(is.null(loadHidecanExample()) & 
+       is.null(input_hidecan()) & 
+       is.null(loadViewpoly())){
+      warning("Select one of the options in `upload` session")
+      return(NULL)
+    } else if(!is.null(loadViewpoly())){
+      return(loadViewpoly()$hidecan)
+    } else if(!is.null(input_hidecan())){
+      return(input_hidecan())
+    } else if(!is.null(loadHidecanExample())){
+      return(loadHidecanExample())
+    }
+  })
+  
   observe({
     if (!is.null(loadMap()) | !is.null(loadQTL())) {
       Sys.sleep(1)
@@ -959,6 +1103,7 @@ mod_upload_server <- function(input, output, session, parent_session){
                               vcf = NULL,
                               align = NULL,
                               wig = NULL,
+                              hidecan = loadHidecan(),
                               version = packageVersion("viewpoly")), 
                          class = "viewpoly")
         assign(input$data.name, obj)
@@ -974,7 +1119,8 @@ mod_upload_server <- function(input, output, session, parent_session){
               loadJBrowse_gff3 = reactive(loadJBrowse_gff3()), 
               loadJBrowse_vcf = reactive(loadJBrowse_vcf()),
               loadJBrowse_align = reactive(loadJBrowse_align()),
-              loadJBrowse_wig = reactive(loadJBrowse_wig())))
+              loadJBrowse_wig = reactive(loadJBrowse_wig()),
+              loadHidecan = reactive(loadHidecan())))
 }
 
 ## To be copied in the UI